@@ -16,9 +16,9 @@
     "@syncfusion/ej2-react-schedule": "^19.4.50",
     "jwt-decode": "^4.0.0",
     "lucide-react": "^0.475.0",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.10.9",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.54.2",
     "react-icons": "^4.3.1",
     "react-quill": "^2.0.0-beta.4",
